Enable converting all the .tsv manifests for cmsc
